+++ include/nil/crypto3/zk/commitments/detail/polynomial/basic_batched_fri.hpp
@@ -30,6 +30,7 @@
 #include <nil/crypto3/marshalling/algebra/types/field_element.hpp>
 
 #include <nil/crypto3/math/polynomial/polynomial.hpp>
+#include <nil/crypto3/math/polynomial/polynomial_dfs.hpp>
 #include <nil/crypto3/math/polynomial/lagrange_interpolation.hpp>
 #include <nil/crypto3/math/domains/evaluation_domain.hpp>
 #include <nil/crypto3/math/algorithms/make_evaluation_domain.hpp>
@@ -49,11 +50,11 @@ namespace nil {
                     /**
                      * @brief Based on the FRI Commitment description from \[ResShift].
                      * @tparam d ...
-                     * @tparam Rounds Denoted by r in \[RedShift].
+                     * @tparam Rounds Denoted by r in \[Placeholder].
                      *
                      * References:
-                     * \[RedShift]:
-                     * "REDSHIFT: Transparent SNARKs from List
+                     * \[Placeholder]:
+                     * "PLACEHOLDER: Transparent SNARKs from List
                      * Polynomial Commitment IOPs",
                      * Assimakis Kattis, Konstantin Panarin, Alexander Vlasov,
                      * Matter Labs,
@@ -63,7 +64,7 @@ namespace nil {
                              typename MerkleTreeHashType,
                              typename TranscriptHashType,
                              std::size_t M = 2>
-                    struct basic_batched_fri_runtime_size {
+                    struct basic_batched_fri {
 
                         constexpr static const std::size_t m = M;
 
@@ -86,18 +87,20 @@ namespace nil {
             				TranscriptHashType, M>::params_type;
 
                         struct round_proof_type {
-                            bool operator==(const round_proof_type &rhs) const {
-                                return y == rhs.y && p == rhs.p && T_root == rhs.T_root &&
-                                       colinear_value == rhs.colinear_value && colinear_path == rhs.colinear_path;
-                            }
-                            bool operator!=(const round_proof_type &rhs) const {
-                                return !(rhs == *this);
-                            }
+                            // bool operator==(const round_proof_type &rhs) const {
+                            //     return y == rhs.y && p == rhs.p && T_root == rhs.T_root &&
+                            //            colinear_value == rhs.colinear_value && colinear_path == rhs.colinear_path;
+                            // }
+                            // bool operator!=(const round_proof_type &rhs) const {
+                            //     return !(rhs == *this);
+                            // }
+                            // std::array<std::array<typename FieldType::value_type, m>, leaf_size> y;
                             std::vector<std::array<typename FieldType::value_type, m>> y;
                             std::array<merkle_proof_type, m> p;
 
                             typename merkle_tree_type::value_type T_root;
 
+                            // std::array<typename FieldType::value_type, leaf_size> colinear_value;
                             std::vector<typename FieldType::value_type> colinear_value;
                             merkle_proof_type colinear_path;
                         };
@@ -112,25 +115,31 @@ namespace nil {
 
                             std::vector<round_proof_type> round_proofs;    // 0..r-2
 
+                            // std::array<math::polynomial<typename FieldType::value_type>,
+                            //     leaf_size> final_polynomials;
                             std::vector<math::polynomial<typename FieldType::value_type>> final_polynomials;
 
                             commitment_type target_commitment;
                         };
 
-                        static precommitment_type
-                            precommit(const std::vector<math::polynomial_dfs<typename FieldType::value_type>> &poly_dfs,
-                                      const std::shared_ptr<math::evaluation_domain<FieldType>> &D) {
+                        template<typename ContainerType>
+                        static typename std::enable_if<
+                            (std::is_same<typename ContainerType::value_type,
+                                          math::polynomial_dfs<typename FieldType::value_type>>::value),
+                            precommitment_type>::type precommit(
+                                const ContainerType &poly,
+                                const std::shared_ptr<math::evaluation_domain<FieldType>> &D) {
 
-                            std::size_t leaf_size = poly_dfs.size();
+                            std::size_t list_size = poly.size();
                             std::vector<std::vector<std::uint8_t>> y_data;
                             y_data.resize(D->m);
 
                             for (std::size_t i = 0; i < D->m; i++) {
-                                for (std::size_t j = 0; j < leaf_size; j++){
-                                    y_data[i].resize(field_element_type::length()*leaf_size);
-
-                                    field_element_type y_val(poly_dfs[j][i]);
-                                    auto write_iter = y_data[i].begin() + field_element_type::length()*j;
+                                y_data[i].resize(field_element_type::length()*list_size);
+                                for (std::size_t j = 0; j < list_size; j++) {
+                                    
+                                    field_element_type y_val(poly[j][i]);
+                                    auto write_iter = y_data[i].begin() + field_element_type::length() * j;
                                     y_val.write(write_iter, field_element_type::length());
                                 }
                             }
@@ -138,13 +147,17 @@ namespace nil {
                             return precommitment_type(y_data.begin(), y_data.end());
                         }
 
-                        static precommitment_type
-                            precommit(const std::vector<math::polynomial<typename FieldType::value_type>> &poly,
-                                      const std::shared_ptr<math::evaluation_domain<FieldType>> &D) {
+                        template<typename ContainerType>
+                        static typename std::enable_if<
+                            (std::is_same<typename ContainerType::value_type,
+                                          math::polynomial<typename FieldType::value_type>>::value),
+                            precommitment_type>::type precommit(
+                                const ContainerType &poly,
+                                const std::shared_ptr<math::evaluation_domain<FieldType>> &D) {
 
-                            std::size_t leaf_size = poly.size();
-                            std::vector<math::polynomial_dfs<typename FieldType::value_type>> poly_dfs(leaf_size);
-                            for (std::size_t i = 0; i < leaf_size; i++){
+                            std::size_t list_size = poly.size();
+                            std::vector<math::polynomial_dfs<typename FieldType::value_type>> poly_dfs(list_size);
+                            for (std::size_t i = 0; i < list_size; i++) {
                                 poly_dfs[i].from_coefficients(poly[i]);
                             }
 
@@ -155,30 +168,140 @@ namespace nil {
                             return P.root();
                         }
 
-                        template <typename PolynomialType>
-                        static commitment_type commit(const std::vector<PolynomialType> &poly,
+                        template <typename ContainerType>
+                        static commitment_type commit(const ContainerType &poly,
                                                       const std::shared_ptr<math::evaluation_domain<FieldType>> &D) {
                             return commit(precommit(poly, D));
                         }
 
-                        static proof_type proof_eval(const std::vector<math::polynomial<typename FieldType::value_type>> Q,
-                                                     const std::vector<math::polynomial<typename FieldType::value_type>> g,
-                                                     precommitment_type &T,
-                                                     const params_type &fri_params,
-                                                     transcript_type &transcript = transcript_type()) {
+                        template <typename ContainerType>
+                        static typename std::enable_if<
+                            (std::is_same<typename ContainerType::value_type,
+                                          math::polynomial_dfs<typename FieldType::value_type>>::value),
+                            proof_type>::type
+                            proof_eval(const ContainerType &Q,
+                                       const ContainerType &g,
+                                       precommitment_type &T,
+                                       const params_type &fri_params,
+                                       transcript_type &transcript = transcript_type()) {
 
                             assert(Q.size() == g.size());
                             std::size_t leaf_size = Q.size();
 
                             proof_type proof;
 
-                            std::vector<math::polynomial<typename FieldType::value_type>> f = Q;    // copy?
+                            ContainerType f = Q;    // copy?
 
                             transcript(commit(T));
 
                             // TODO: how to sample x?
                             std::size_t domain_size = fri_params.D[0]->m;
-                            std::uint64_t x_index = (transcript.template int_challenge<std::size_t>())%domain_size;
+                            std::uint64_t x_index = (transcript.template int_challenge<std::uint64_t>()) % domain_size;
+
+                            std::size_t r = fri_params.r;
+
+                            std::vector<round_proof_type> round_proofs;
+                            std::unique_ptr<merkle_tree_type> p_tree = std::make_unique<merkle_tree_type>(T);
+                            merkle_tree_type T_next;
+
+                            for (std::size_t i = 0; i < r - 1; i++) {
+
+                                std::size_t domain_size = fri_params.D[i]->m;
+
+                                typename FieldType::value_type alpha = transcript.template challenge<FieldType>();
+
+                                x_index %= domain_size;
+
+                                // m = 2, so:
+                                std::array<std::size_t, m> s_indices;
+                                if constexpr (m == 2) {
+                                    s_indices[0] = x_index;
+                                    s_indices[1] = (x_index + domain_size / 2) % domain_size;
+                                } else {
+                                    return {};
+                                }
+
+
+                                // std::array<std::array<typename FieldType::value_type, m>, leaf_size> y;
+                                std::vector<std::array<typename FieldType::value_type, m>> y(leaf_size);
+
+                                for (std::size_t polynom_index = 0; polynom_index < leaf_size; polynom_index++) {
+                                    for (std::size_t j = 0; j < m; j++) {
+                                        y[polynom_index][j] =
+                                            (i == 0 ? g[polynom_index][s_indices[j]] :
+                                                      f[polynom_index][s_indices[j]]);    // polynomial evaluation
+                                    }
+                                }
+
+                                std::array<merkle_proof_type, m> p;
+
+                                for (std::size_t j = 0; j < m; j++) {
+
+                                    p[j] = merkle_proof_type(*p_tree, s_indices[j]);
+                                }
+
+                                // std::array<typename FieldType::value_type, leaf_size> colinear_value;
+                                std::vector<typename FieldType::value_type> colinear_value(leaf_size);
+
+                                for (std::size_t polynom_index = 0; polynom_index < leaf_size; polynom_index++) {
+                                    f[polynom_index] =
+                                        fold_polynomial<FieldType>(f[polynom_index], alpha, fri_params.D[i]);
+                                }
+
+                                x_index = x_index % (fri_params.D[i + 1]->m);
+
+                                for (std::size_t polynom_index = 0; polynom_index < leaf_size; polynom_index++) {
+                                    colinear_value[polynom_index] =
+                                        f[polynom_index][x_index];    // polynomial evaluation
+                                }
+
+                                T_next = precommit(f, fri_params.D[i + 1]);    // new merkle tree
+                                transcript(commit(T_next));
+
+                                merkle_proof_type colinear_path = merkle_proof_type(T_next, x_index);
+
+                                round_proofs.push_back(
+                                    round_proof_type({y, p, p_tree->root(), colinear_value, colinear_path}));
+
+                                p_tree = std::make_unique<merkle_tree_type>(T_next);
+                            }
+
+                            std::vector<math::polynomial<typename FieldType::value_type>>
+                                final_polynomials(f.size());
+                                
+                            for (std::size_t polynom_index = 0; polynom_index < f.size(); polynom_index++){
+                                final_polynomials[polynom_index] =
+                                    math::polynomial<typename FieldType::value_type>(
+                                        f[polynom_index].coefficients());
+                            }
+
+                            return proof_type({round_proofs, final_polynomials, commit(T)});
+                        }
+
+                        template <typename ContainerType>
+                        static typename std::enable_if<
+                            (std::is_same<typename ContainerType::value_type,
+                                          math::polynomial<typename FieldType::value_type>>::value),
+                            proof_type>::type
+                            proof_eval(const ContainerType &Q,
+                                       const ContainerType &g,
+                                       precommitment_type &T,
+                                       const params_type &fri_params,
+                                       transcript_type &transcript = transcript_type()) {
+
+                            assert(Q.size() == g.size());
+                            std::size_t leaf_size = Q.size();
+
+                            proof_type proof;
+
+                            ContainerType f = Q;    // copy?
+
+                            transcript(commit(T));
+
+                            // TODO: how to sample x?
+                            std::size_t domain_size = fri_params.D[0]->m;
+                            std::uint64_t x_index = (transcript.template int_challenge<std::uint64_t>()) % domain_size;
+
                             typename FieldType::value_type x = fri_params.D[0]->get_domain_element(x_index);
 
                             std::size_t r = fri_params.r;
@@ -193,20 +316,7 @@ namespace nil {
 
                                 typename FieldType::value_type alpha = transcript.template challenge<FieldType>();
 
-                                std::vector<math::polynomial<typename FieldType::value_type>> f_next(leaf_size);
-
-                                for (std::size_t polynom_index = 0; polynom_index < leaf_size;
-                                            polynom_index++){
-                                    f_next[polynom_index] = fold_polynomial<FieldType>(
-                                        f[polynom_index], alpha);
-                                }
-
                                 x_index %= domain_size;
-
-                                std::uint64_t x_next_index = x_index % (fri_params.D[i + 1]->m);
-
-                                typename FieldType::value_type x_next =
-                                    fri_params.D[i+1]->get_domain_element(x_next_index);
 
                                 // m = 2, so:
                                 std::array<typename FieldType::value_type, m> s;
@@ -215,76 +325,84 @@ namespace nil {
                                     s[0] = x;
                                     s[1] = -x;
                                     s_indices[0] = x_index;
-                                    s_indices[1] = (x_index + domain_size/2)%domain_size;
+                                    s_indices[1] = (x_index + domain_size / 2) % domain_size;
                                 } else {
                                     return {};
                                 }
 
+
+                                // std::array<std::array<typename FieldType::value_type, m>, leaf_size> y;
                                 std::vector<std::array<typename FieldType::value_type, m>> y(leaf_size);
 
-                                for (std::size_t polynom_index = 0; polynom_index < leaf_size;
-                                            polynom_index++){
+                                for (std::size_t polynom_index = 0; polynom_index < leaf_size; polynom_index++) {
                                     for (std::size_t j = 0; j < m; j++) {
                                         y[polynom_index][j] =
                                             (i == 0 ? g[polynom_index].evaluate(s[j]) :
-                                            f[polynom_index].evaluate(s[j]));    // polynomial evaluation
+                                                      f[polynom_index].evaluate(s[j]));    // polynomial evaluation
                                     }
                                 }
 
                                 std::array<merkle_proof_type, m> p;
 
                                 for (std::size_t j = 0; j < m; j++) {
+
                                     p[j] = merkle_proof_type(*p_tree, s_indices[j]);
                                 }
 
+                                // std::array<typename FieldType::value_type, leaf_size> colinear_value;
                                 std::vector<typename FieldType::value_type> colinear_value(leaf_size);
 
-                                for (std::size_t polynom_index = 0; polynom_index < leaf_size;
-                                            polynom_index++){
-                                    colinear_value[polynom_index] =
-                                        f_next[polynom_index].evaluate(x_next);    // polynomial evaluation
+                                for (std::size_t polynom_index = 0; polynom_index < leaf_size; polynom_index++) {
+                                    f[polynom_index] = fold_polynomial<FieldType>(f[polynom_index], alpha);
                                 }
 
-                                T_next = precommit(f_next, fri_params.D[i + 1]);    // new merkle tree
+                                x_index = x_index % (fri_params.D[i + 1]->m);
+                                x = fri_params.D[i+1]->get_domain_element(x_index);
+
+                                for (std::size_t polynom_index = 0; polynom_index < leaf_size; polynom_index++) {
+                                    colinear_value[polynom_index] =
+                                        f[polynom_index].evaluate(x);    // polynomial evaluation
+                                }
+
+                                T_next = precommit(f, fri_params.D[i + 1]);    // new merkle tree
                                 transcript(commit(T_next));
 
-                                merkle_proof_type colinear_path = merkle_proof_type(T_next, x_next_index);
+                                merkle_proof_type colinear_path = merkle_proof_type(T_next, x_index);
 
                                 round_proofs.push_back(
                                     round_proof_type({y, p, p_tree->root(), colinear_value, colinear_path}));
 
                                 p_tree = std::make_unique<merkle_tree_type>(T_next);
-
-                                x = x_next;
-                                x_index = x_next_index;
-
-                                f = f_next;
                             }
-                            return proof_type({round_proofs, f});
+
+                            std::vector<math::polynomial<typename FieldType::value_type>>
+                                final_polynomials(f.begin(), f.end());
+
+                            return proof_type({round_proofs, final_polynomials, commit(T)});
                         }
 
-                        static bool verify_eval(proof_type &proof,
-                                                params_type &fri_params,
-                                                const std::vector<
-                                                    math::polynomial<typename FieldType::value_type>> U,
-                                                const std::vector<
-                                                    math::polynomial<typename FieldType::value_type>> V,
-                                                transcript_type &transcript = transcript_type()) {
-
-                            transcript(proof.target_commitment);
+                        template <typename ContainerType>
+                        static bool
+                            verify_eval(proof_type &proof,
+                                        params_type &fri_params,
+                                        const ContainerType U,
+                                        const ContainerType V,
+                                        transcript_type &transcript = transcript_type()) {
 
                             assert(U.size() == V.size());
                             std::size_t leaf_size = U.size();
 
-                            std::size_t idx = transcript.template int_challenge<std::size_t>();
+                            transcript(proof.target_commitment);
+
+                            std::uint64_t idx = transcript.template int_challenge<std::uint64_t>();
                             typename FieldType::value_type x = fri_params.D[0]->get_domain_element(idx);
 
                             std::size_t r = fri_params.r;
 
-                            for (std::size_t i = 0; i < r; i++) {
+                            for (std::size_t i = 0; i < r - 1; i++) {
                                 typename FieldType::value_type alpha = transcript.template challenge<FieldType>();
 
-                                typename FieldType::value_type x_next = x*x;
+                                typename FieldType::value_type x_next = x * x;
 
                                 // m = 2, so:
                                 std::array<typename FieldType::value_type, m> s;
@@ -296,16 +414,15 @@ namespace nil {
                                 }
 
                                 for (std::size_t j = 0; j < m; j++) {
-                                    std::vector<std::uint8_t> leaf_data (field_element_type::length() * leaf_size);
+                                    std::vector<std::uint8_t> leaf_data(field_element_type::length() * leaf_size);
 
-                                    for (std::size_t polynom_index = 0; polynom_index < leaf_size;
-                                        polynom_index++){
+                                    for (std::size_t polynom_index = 0; polynom_index < leaf_size; polynom_index++) {
 
                                         typename FieldType::value_type leaf = proof.round_proofs[i].y[polynom_index][j];
 
                                         field_element_type leaf_val(leaf);
-                                        auto write_iter = leaf_data.begin() +
-                                            field_element_type::length() * polynom_index;
+                                        auto write_iter =
+                                            leaf_data.begin() + field_element_type::length() * polynom_index;
                                         leaf_val.write(write_iter, field_element_type::length());
                                     }
 
@@ -314,24 +431,24 @@ namespace nil {
                                     }
                                 }
 
-                                std::vector<std::uint8_t> leaf_data (field_element_type::length() * leaf_size);
+                                std::vector<std::uint8_t> leaf_data(field_element_type::length() * leaf_size);
 
-                                for (std::size_t polynom_index = 0; polynom_index < leaf_size;
-                                    polynom_index++){
+                                for (std::size_t polynom_index = 0; polynom_index < leaf_size; polynom_index++) {
 
                                     std::array<typename FieldType::value_type, m> y;
 
                                     for (std::size_t j = 0; j < m; j++) {
                                         if (i == 0) {
-                                            y[j] = 
-                                                (proof.round_proofs[i].y[polynom_index][j] - 
-                                                    U[polynom_index].evaluate(s[j])) / V[polynom_index].evaluate(s[j]);
+                                            y[j] = (proof.round_proofs[i].y[polynom_index][j] -
+                                                    U[polynom_index].evaluate(s[j])) /
+                                                   V[polynom_index].evaluate(s[j]);
                                         } else {
                                             y[j] = proof.round_proofs[i].y[polynom_index][j];
                                         }
                                     }
 
-                                    std::vector<std::pair<typename FieldType::value_type, typename FieldType::value_type>>
+                                    std::vector<
+                                        std::pair<typename FieldType::value_type, typename FieldType::value_type>>
                                         interpolation_points {
                                             std::make_pair(s[0], y[0]),
                                             std::make_pair(s[1], y[1]),
@@ -340,13 +457,15 @@ namespace nil {
                                     math::polynomial<typename FieldType::value_type> interpolant =
                                         math::lagrange_interpolation(interpolation_points);
 
-                                    typename FieldType::value_type leaf = proof.round_proofs[i].colinear_value[polynom_index];
-                                    
+                                    typename FieldType::value_type leaf =
+                                        proof.round_proofs[i].colinear_value[polynom_index];
+
                                     field_element_type leaf_val(leaf);
                                     auto write_iter = leaf_data.begin() + field_element_type::length() * polynom_index;
                                     leaf_val.write(write_iter, field_element_type::length());
 
-                                    if (interpolant.evaluate(alpha) != proof.round_proofs[i].colinear_value[polynom_index]) {
+                                    if (interpolant.evaluate(alpha) !=
+                                        proof.round_proofs[i].colinear_value[polynom_index]) {
                                         return false;
                                     }
                                 }
@@ -363,8 +482,7 @@ namespace nil {
                             //     return false;
                             // }
 
-                            for (std::size_t polynom_index = 0; polynom_index < leaf_size;
-                                    polynom_index++){
+                            for (std::size_t polynom_index = 0; polynom_index < leaf_size; polynom_index++) {
 
                                 if (proof.final_polynomials[polynom_index].degree() >
                                     std::pow(2, std::log2(fri_params.max_degree + 1) - r + 1) - 1) {
@@ -376,23 +494,23 @@ namespace nil {
                             return true;
                         }
 
-                        static bool verify_eval(proof_type &proof,
-                                                params_type &fri_params,
-                                                const std::vector<
-                                                    math::polynomial<typename FieldType::value_type>> U,
-                                                const math::polynomial<typename FieldType::value_type> V,
-                                                transcript_type &transcript = transcript_type()) {
-
-                            transcript(proof.target_commitment);
+                        template <typename ContainerType>
+                        static bool
+                            verify_eval(proof_type &proof,
+                                        params_type &fri_params,
+                                        const ContainerType U,
+                                        const math::polynomial<typename FieldType::value_type> V,
+                                        transcript_type &transcript = transcript_type()) {
 
                             std::size_t leaf_size = U.size();
+                            transcript(proof.target_commitment);
 
-                            std::size_t idx = transcript.template int_challenge<std::size_t>();
+                            std::uint64_t idx = transcript.template int_challenge<std::uint64_t>();
                             typename FieldType::value_type x = fri_params.D[0]->get_domain_element(idx);
 
                             std::size_t r = fri_params.r;
 
-                            for (std::size_t i = 0; i < r; i++) {
+                            for (std::size_t i = 0; i < r - 1; i++) {
                                 typename FieldType::value_type alpha = transcript.template challenge<FieldType>();
 
                                 typename FieldType::value_type x_next = x*x;
@@ -407,16 +525,15 @@ namespace nil {
                                 }
 
                                 for (std::size_t j = 0; j < m; j++) {
-                                    std::vector<std::uint8_t> leaf_data (field_element_type::length() * leaf_size);
+                                    std::vector<std::uint8_t> leaf_data(field_element_type::length() * leaf_size);
 
-                                    for (std::size_t polynom_index = 0; polynom_index < leaf_size;
-                                        polynom_index++){
+                                    for (std::size_t polynom_index = 0; polynom_index < leaf_size; polynom_index++) {
 
                                         typename FieldType::value_type leaf = proof.round_proofs[i].y[polynom_index][j];
 
                                         field_element_type leaf_val(leaf);
-                                        auto write_iter = leaf_data.begin() +
-                                            field_element_type::length() * polynom_index;
+                                        auto write_iter =
+                                            leaf_data.begin() + field_element_type::length() * polynom_index;
                                         leaf_val.write(write_iter, field_element_type::length());
                                     }
 
@@ -425,25 +542,24 @@ namespace nil {
                                     }
                                 }
 
-                                std::vector<std::uint8_t> leaf_data (field_element_type::length() * leaf_size);
+                                std::vector<std::uint8_t> leaf_data(field_element_type::length() * leaf_size);
 
-                                for (std::size_t polynom_index = 0; polynom_index < leaf_size;
-                                    polynom_index++){
+                                for (std::size_t polynom_index = 0; polynom_index < leaf_size; polynom_index++) {
 
                                     std::array<typename FieldType::value_type, m> y;
 
                                     for (std::size_t j = 0; j < m; j++) {
                                         if (i == 0) {
-                                            y[j] = 
-                                                (proof.round_proofs[i].y[polynom_index][j] - 
-                                                    U[polynom_index].evaluate(s[j])) / V.evaluate(s[j]);
+                                            y[j] = (proof.round_proofs[i].y[polynom_index][j] -
+                                                    U[polynom_index].evaluate(s[j])) /
+                                                   V.evaluate(s[j]);
                                         } else {
                                             y[j] = proof.round_proofs[i].y[polynom_index][j];
                                         }
                                     }
-                                
 
-                                    std::vector<std::pair<typename FieldType::value_type, typename FieldType::value_type>>
+                                    std::vector<
+                                        std::pair<typename FieldType::value_type, typename FieldType::value_type>>
                                         interpolation_points {
                                             std::make_pair(s[0], y[0]),
                                             std::make_pair(s[1], y[1]),
@@ -452,14 +568,15 @@ namespace nil {
                                     math::polynomial<typename FieldType::value_type> interpolant =
                                         math::lagrange_interpolation(interpolation_points);
 
-                                    typename FieldType::value_type leaf = proof.round_proofs[i].colinear_value[polynom_index];
+                                    typename FieldType::value_type leaf =
+                                        proof.round_proofs[i].colinear_value[polynom_index];
 
-                                    
                                     field_element_type leaf_val(leaf);
                                     auto write_iter = leaf_data.begin() + field_element_type::length() * polynom_index;
                                     leaf_val.write(write_iter, field_element_type::length());
 
-                                    if (interpolant.evaluate(alpha) != proof.round_proofs[i].colinear_value[polynom_index]) {
+                                    if (interpolant.evaluate(alpha) !=
+                                        proof.round_proofs[i].colinear_value[polynom_index]) {
                                         return false;
                                     }
                                 }
@@ -468,6 +585,7 @@ namespace nil {
                                 if (!proof.round_proofs[i].colinear_path.validate(leaf_data)) {
                                     return false;
                                 }
+
                                 x = x_next;
                             }
 
@@ -476,8 +594,7 @@ namespace nil {
                             //     return false;
                             // }
 
-                            for (std::size_t polynom_index = 0; polynom_index < leaf_size;
-                                    polynom_index++){
+                            for (std::size_t polynom_index = 0; polynom_index < leaf_size; polynom_index++) {
 
                                 if (proof.final_polynomials[polynom_index].degree() >
                                     std::pow(2, std::log2(fri_params.max_degree + 1) - r + 1) - 1) {
